feat(filter): Filter current tea element on the page - server and client implementation
Currently the site can support the feature to filter tea listing on a any
page. You can now add a css class called .filterable to any HTML object,
xxx they have data attributes of teaid then it will request the server to
filter items on screen. The JavaScript updates includes the ability to
look for filterable DOM elements and generate POST scripts to the server.
I am also parsing and templating JSON data that is served back from our
backend. Further work will be required to extend this to search, it only
works on /teas/catalog. However, adding the filterable tag and data
attributes is the minimal effort needed to enable this feature.

✔️ views.py - most_popular(), origin(), type() new methods - very similar
functionality, either returning a full list from the database or filtering
specific teas based on PKs. The models are convered to JSON objects and
sent back to the client.

✔️ urls.py - uncommented type and origin url patterns.

✔️ tea_catalog.html - added new filter buttons to the dropdown and included
data attributes to the list-group element on the page. These were mainly
changed for testing the filter, changes to the overall feel is to be
expected.

✔️ main.js - findTeaIds() responsible for looping through the list of HTML
elements with the .filterable css classname. Ajax to make POST requests to the server in order to get data
back for filtering.

Removed catalog.html in favour of tea_catalog.html and avoid duplication