feat(save tea): users save tea button functionality created
New feature - saving a tea to a user. Also, new function inside the specific tea
code which starts incrementing page views.

✔️ views.py - add_favourite_tea() and increment_tea_page_views() are both
new. add_favourite_tea() handles saving the current user and tea requested
to the SavedTea model. Increment_tea_page_views() is responsible for
incrementing page views to the database when ever the specific_tea()
method is called.

✔️ main.js - event handler for save tea button click and AJAX get request
which sends the ID of the tea saved

✔️ specific_tea.html - dumped some test html for save buttons

✔️ urls.py - added save to the url mapping